First pass at gradient based scaling. This affected many files. The gradient based scaling needs the NLP starting point and I made some minor changes to the NLP interface to allow NULL for some of the entries in GetStartingPoint. I also added a StandardScalingBase class for all scaling methods that will use the standard ScalingMatrices. All the derived classes need to do is calculate the scaling parameters - already, much of the code between UserScaling and GradientBasedScaling is then shared.
I continued to work on the scaling and added unscaled_f,c,d methods
to the IpoptNLP. These methods have also been added to the 
Calc Quant. The values are cached in the IpOrig, but not the 
IpResto. If they are cached in the IpoptNLP and not the CalcQuant.
it would remove our depends_on_mu problem.