Changes for efficiency in the Restoration phase. Some of the changes improve the efficiency of the regular algorithm as well.
Changed the order of ifs in the FilterConvCheck to
avoid unnecessary calcs. Minor changes in IpCq.
Wrapped many journalist calls in the iteration
outputs (with if (Jnlst().ProduceOutput...))
Made the CompoundMatrix methods more efficient
when all the matrix blocks are on the diagonal 
(the matrix space can now easily indicate if
this is so).