fix multipliers for fixed variables if fixed_variable_treatment is make_constraint and solve failed
At least when the NLP solve returned with status RESTORATION_FAILURE, the vector y_c is homogeneously zero.
In this case, setting the entries in full_z_L|U for the fixed variables failed, since the implementation
assumes (and asserts) that y_c is not a homogeneous vector.
This commit works around a segmentation fault in optimized mode by adding an implementation for the case
that y_c is homogeneous.