by Toni Kelman: automate some steps for building mex files on Mac and Windows
This adds get.Gnumex to /Ipopt/contrib/MatlabInterface along with a patch
I wrote to make Gnumex work on Win64 Matlab.

Gnumex doesn't need to build any executables, it runs from inside
Matlab. It comes with some mexw32 files, and part of my patch tries to
build the mexw64 equivalents from inside Matlab (if 64 bit), or uses a
different workaround if that fails. I added a gnumex target in
Makefile.in to open an instance of Matlab and run Gnumex (or echo a
warning with instructions if the user hasn't run get.Gnumex yet) with
some of the options filled out to recommended settings. If Cygwin and/or
MinGW are installed in default locations, the user should only need to
click "Make options file" then exit the new instance of Matlab (so we
know Gnumex is finished).

The new mexopts target in Makefile.in makes the Mac or Windows
modifications to mexopts.sh or mexopts.bat. I added it as a recursive
make call in the $(TARGET) rule, since adding it to the dependencies
would make "mexopts" show up on the compilation command line from "$^".
We could alternately have "make mexopts" always be a manual step for Mac
and Windows users.

For Mac, if there is not currently a mexopts.sh file in the Matlab
interface build directory, I look at the default one, test if the
hardcoded $SDKROOT path actually exists, and if not I make a copy with
all references to $SDKROOT removed.

For Windows, I give a warning if no mexopts.bat file is found in the
Matlab interface build directory. This warning message may take a lot of
scrolling to find unfortunately, if the default setup results in really
long error messages from mex. If there is a mexopts.bat in the build
directory, I add $(LIBS) to GM_ADD_LIBS, change COMPILER=gcc to
COMPILER=g++, and GM_MEXLANG=c to GM_MEXLANG=cxx. I need to modify -L
flags from $(LIBS) to Windows format directories, although forward
slashes work and prevent having to double/quadruple/octuple backslashes
for future sed calls. I do this using "cygpath -m" on $(libdir) and
/usr/lib in Cygwin, or using a fairly ugly cmd /c DOS for-do-@echo call
in MinGW (only on libdir).

An annoyance in Cygwin is that due to the alternatives system, g++ is a
symbolic link to either g++-4 or g++-3. This causes "Access is denied"
errors when linking mex files, so on Cygwin I use g++-4 for COMPILER, or
g++-3 if the former is not on the path, or g++ if neither g++-4 or g++-3
is on the path.

Results:
Linux - no change, except more output to the screen when going through the no-op make mexopts.

Mac - working mex file when configured with --disable-shared, but
crashes Matlab without. The shared-library version gives a malloc error,
"pointer being freed was not allocated," in
IpoptOptions::useQuasiNewton, Ipopt::OptionsList::GetStringValue,
std::string::assign. This is probably a libstdc++ incompatibility issue.

MinGW32 - works very nicely with no configure flags needed (../configure
&& make install test && cd Ipopt/contrib/MatlabInterface/src && make
gnumex install && echo done)

MinGW64 - works with the configure flag ADD_FFLAGS=-static-libgcc, this
prevents a "multiple definition of `_Unwind_SjLj_Unregister'" error when
using the TDM-GCC distribution of MinGW64.

Cygwin - compiles mex file successfully, but freezes Matlab when
configured with --disable-shared, crashes Matlab without. The
freeze/crash is extremely difficult to debug, I couldn't even get any
output by adding a printf at the very beginning of mexFunction in
ipopt.cpp.