Additional checks of initial point input and callback function outputs - Error if initial point is complex or sparse - Error if any callback returns complex value - Error if gradient or constraint does not return a double - Convert sparse objective, gradient (done in R2133), or constraint to full - Convert non-double Jacobian structure, Hessian structure, Jacobian, or Hessian to double
This fixes a segfault on large problems if structure callbacks return Matlab logical datatypes (A ~= 0), the line copymemory(mxGetPr(ptr),x,nnz) in the SparseMatrix constructor was going past the valid range of the mxGetPr(ptr) array since logicals are smaller than doubles. This was still giving correct behavior for problems with few enough nonzeros not to segfault, since only the row and column indices are used from the structure matrices and it didn't matter that the values contained garbage.