fix to ipoptr (contributed by Anirban Mukherjee and Jelmer Ypma)
Jelmer suggested that I try changing these lines in R/ipoptr.R

# build ipoptr object
    ret <- list( "x0"=x0,
                 "eval_f"=eval_f_wrapper,
                 "eval_grad_f"=eval_grad_f_wrapper,
                 "lower_bounds"=lb,
                 "upper_bounds"=ub,
                 "eval_g"=eval_g_wrapper,
                 "eval_jac_g"=eval_jac_g_wrapper,
                 "constraint_lower_bounds"=constraint_lb,
                 "constraint_upper_bounds"=constraint_ub,
                 "eval_jac_g_structure"=eval_jac_g_structure,
                 "eval_h"=eval_h_wrapper,
                 "eval_h_structure"=eval_h_structure,
                 "options"=get.option.types(opts) )

to

# build ipoptr object
    ret <- list( "x0"=x0,
                 "eval_f"=eval_f_wrapper,
                 "eval_grad_f"=eval_grad_f_wrapper,
                 "lower_bounds"=lb,
                 "upper_bounds"=ub,
                 "eval_g"=eval_g_wrapper,
                 "eval_jac_g"=eval_jac_g_wrapper,
                 "constraint_lower_bounds"=constraint_lb,
                 "constraint_upper_bounds"=constraint_ub,
                 "eval_jac_g_structure"=eval_jac_g_structure,
                 "eval_h"=eval_h_wrapper,
                 "eval_h_structure"=eval_h_structure,
                 "options"=get.option.types(opts),

                 "environment" = new.env() )

In ipopt.r I see one additional element in the ret list
("ipoptr_environment"=ipoptr_environment). I first left that in and
added the element/argument suggested ("environment" = new.env()) to get:

    # build ipoptr object
    ret <- list( "x0"=x0,
                 "eval_f"=eval_f_wrapper,
                 "eval_grad_f"=eval_grad_f_wrapper,
                 "lower_bounds"=lb,
                 "upper_bounds"=ub,
                 "eval_g"=eval_g_wrapper,
                 "eval_jac_g"=eval_jac_g_wrapper,
                 "constraint_lower_bounds"=constraint_lb,
                 "constraint_upper_bounds"=constraint_ub,
                 "eval_jac_g_structure"=eval_jac_g_structure,
                 "eval_h"=eval_h_wrapper,
                 "eval_h_structure"=eval_h_structure,
                 "options"=get.option.types(opts),
                 "ipoptr_environment"=ipoptr_environment,
                  "environment" = new.env() )

Installing this from R as a package worked as expected.

However, revision 8 on R-forge says that Jelmer removed
ipoptr_environment because of a conflict with data.table. (I am not sure
why ipopt 3.11.3, which I downloaded last week, still has this argument
given Jelmer's revision dated March 2012.)

https://r-forge.r-project.org/scm/viewvc.php/pkg/R/ipoptr.R?view=markup&root=ipoptr

I tried removing the ipoptr_environment argument but got the following
error message:

Error: is.environment(x$ipoptr_environment) is not TRUE

It did not seem that ipoptr_environment was doing much. So I went into
is.ipopt.R and changed line 18 which checks if "ipoptr_environment" is
an R environment, to a check if "environment" is an environment.

I changed

stopifnot( is.environment(x$ipoptr_environment) )

to

stopifnot( is.environment(x$environment) )

This works as expected on my system. Ipoptr loads and gives the right
solution in the examples.

Summary: in ipopt.R I commented out the argument
("ipoptr_environment"=ipoptr_environment) and added the argument
("environment" = new.env() ). In is.ipopt.R I changed line 18 which
checks if ipoptr_environment is an environment to a check for
environment being an environment. For redundancy I am attaching both .R
files to this email.