apply changes from pull request !5: support custom algorithm development
See https://github.com/coin-or/Ipopt/pull/5:
This pull request mainly consists of a reorganization of the AlgorithmBuilder class to allow easier customization of the Ipopt algorithm. In particular, we wanted to make use of the code that creates the SymLinearSolver object to implement our own SymLinearSolver without copy-pasting everything in AlgorithmBuilder.
AlgorithmBuilder::BuildBasicAlgorithm now consists of 8 method calls that build the core components passed into the arguments of the IpoptAlgorithm class. These calls are ordered based on any dependencies they might have. In addition, all code for creating the PDSystemSolver, AugSystemSolver, and SymLinearSolver has been moved into separate factory methods.
Also, included is a change to install a few more header files with Ipopt. Some of these are required to subclass AlgorithmBuilder, and the others are simply some matrix types that we require.