change default for ma97_print_level to -1
Some e-mail exchange on this topic:

On 11/12/2015 11:48 AM, Jonathan Hogg wrote:
> The system solved by MA97 changes each iteration, with diagonal entries
> in the (1,1) block becoming more extreme towards 0 and +infinity [give
> or take the diagonal entries of the Hessian] depending on whether the
> associated constraint is active or not. This is what I'd expect to be
> causing the singular error messages at some iterations and not at others
> - it is perfectly normal for some problems (and by singular here we mean
> effectively numerically singular rather than structurally or in exact
> arithmetic).
>
> IIRC Ipopt just takes a singular error return as a signal to increase
> regularization of the system in the (1,1) block and then tries again.
> Hence it is probably just a case of needing a patch to tell MA97 not to
> print error messages where the user can see them :P.

Stefan Vigerske, 18 December 2015:
I guess we could suppress all MA97 messages in Ipopt by changing
ma97_print_level to a negative number. Currently it defaults to 0, which
means "error and warning messages only".
I'm not sure that disabling error messages by default is a good thing to do.
If a numerical singular system is not a really serious error, couldn't
one change the message to appear only on print level 1 ("limited
diagnostic printing") or higher?

Jonathan Hogg:
The thing is, it /is/ a serious error for the linear solver - it can't
really guarantee much about the quality of the solution it will deliver.
So it needs to stay as a printed error by default or most users of the code.
However as Ipopt detects the error and fixes it by regularizing the
problem, it's not a serious error as far as Ipopt is concerned.
I think the solution of setting the ma97 print level to off and reporting
any "unhandled" error returns via an Ipopt error message is probably the
best way to go.